+++ gDRplots/NEWS.md
@@ -1,3 +1,8 @@
+## gDRplots 0.0.90 - 2025-07-31
+* add rounding options to `generate_datatable`
+* add functions to create associations summary tables for PRISM
+* add associations test data
+
 ## gDRplots 0.0.89 - 2025-07-29
 * update var names